Move toPlainObject.ts to util; create testUtil/spyToPlainObject.ts.
