+++ packages/common/src/index.ts
@@ -50,7 +50,7 @@ export * from "./types/SpokenForm";
 export * from "./util/textFormatters";
 export * from "./types/snippet.types";
 export * from "./testUtil/fromPlainObject";
-export * from "./testUtil/toPlainObject";
+export * from "./testUtil/spyToPlainObject";
 export { default as DefaultMap } from "./util/DefaultMap";
 export * from "./types/GeneralizedRange";
 export * from "./types/RangeOffsets";
@@ -92,3 +92,4 @@ export * from "./getFakeCommandServerApi";
 export * from "./types/TestCaseFixture";
 export * from "./util/getEnvironmentVariableStrict";
 export * from "./util/CompositeKeyDefaultMap";
+export * from "./util/toPlainObject";
